Внес правки и добавил новый файл с функциями
